requirements.txt
aiogram==2.23.1
python-dotenv==0.21.1
Django==4.1.3

loguru==0.7.2